refactor(settings): extract agent-home skill import
